Fix frame limiting, fix double updates for moved objects
Now the program stops at the desired amount of frames
and does not overrun the number by 1.
Vector_update previously could process the same object
twice if it was moved during the pass. By itself, that's
not a problem, but I didn't like it because it might
cause issues later...