Implement node recycling (giant commit)
There is a queue of leaf nodes eligible for recycling,
which is checked every frame by tree_optimizeNodes.
It goes through the list and checks if the leaves
can be merged with the other child of their parents.
In that case, objects from one buffer are copied into
the other and the other buffer is assigned to the parent.
The now-empty buffer of one child remains to be used
if the node has to be split again. The parent still
has child pointers, but the objects reside in the parent's
buffer, making naive tree descent functions
very dangerous. Always check the node->buf member to
see if it's a leaf node.